requirements.txt
flask==2.3.2
pydantic==1.10.7
python-dotenv==1.0.0
requests
langchain==0.0.173
pymongo==4.3.3
openai==0.27.6
